+++ CHANGELOG.md
@@ -60,6 +60,7 @@ ImportedLFP().drop()
     - Add `SensorData` to `populate_all_common` #1281
     - Add `fetch1_dataframe` to `SensorData` #1291
     - Allow storage of numpy arrays using `AnalysisNwbfile.add_nwb_object` #1298
+    - `IntervalList.fetch_interval` now returns `Interval` object #1293
 - Position
     - Allow population of missing `PositionIntervalMap` entries during population
         of `DLCPoseEstimation` #1208
@@ -78,13 +79,14 @@ ImportedLFP().drop()
 - Behavior
     - Implement pipeline for keypoint-moseq extraction of behavior syllables #1056
 - LFP
-    - Adding a condition in the MAD detector to replace zero, NaN, or infinite
-        MAD values with 1.0. #1280
+    - Implement `ImportedLFP.make()` for ingestion from nwb files #1278
+    - Adding a condition in the MAD detector to replace zero, NaN, or infinite MAD
+        values with 1.0. #1280
     - Refactoring the creation of LFPElectrodeGroup with added input validation
         and transactional insertion. #1280
     - Updating the LFPBandSelection logic with comprehensive validation and batch
         insertion for electrodes and references. #1280
-    - Implement `ImportedLFP.make()`  for ingestion from nwb files #1278
+    - Implement `ImportedLFP.make()` for ingestion from nwb files #1278
 
 ## [0.5.4] (December 20, 2024)